// src/modules/budget/dto/show-budget.dto.ts
export class ShowBudgetDTO {
    idBudget: number;
    uniqueId: string;
    years: number;
    idGlAccount: number;
    idCostCenter: number;
    value01: number;
    value02: number;
    value03: number;
    value04: number;
    value05: number;
    value06: number;
    value07: number;
    value08: number;
    value09: number;
    value10: number;
    value11: number;
    value12: number;
    value13: number;
    value14: number;
    value15: number;
    value16: number;
    total: number;
    createdAt: Date;
    createdBy: string;
    updatedAt: Date | null;
    updatedBy: string | null;
}